Cleanup int type of range.
Numerical value of range sent by sonar is 0-254 inches. Store in a uint8
instead of uint16 so we can do all range operations in a single machine
instruction on the 8 bit processor. Add -Wall to compiler flags for
sonar string tests.

Testing: unit tests pass. Manual range-led testing passed.

+++ sonar-usart-3led-sleep/sonar-string.c
@@ -44,8 +44,8 @@ void sonar_string_add_char(uint8_t next_char) {
     sonar_string_index++;
 }
 
-uint16_t sonar_string_as_int(uint16_t old_range) {
-    uint16_t new_range = old_range;
+uint8_t sonar_string_as_int(uint8_t old_range) {
+    uint8_t new_range = old_range;
     if (sonar_string_index == sonar_string_length - 1) {
         // last character written should be carriage return to assure
         // that a full string was read....
